# speed up header encryption / decryption